+++ templates/basic/src/mikuru-env.d.ts
@@ -1,0 +1,1 @@
+import "mikuru/env";