+++ templates/basic/vite.config.ts
@@ -1,0 +1,6 @@
+import { defineConfig } from "vite";
+import { mikuru } from "mikuru/vite";
+
+export default defineConfig({
+  plugins: [mikuru()]
+});